going to make PairDisplay and GroupDisplay separate frames
