--- setup.py
@@ -1,39 +1,0 @@
-"""
-This is a setup.py script generated by py2applet
-
-Usage:
-    python setup.py py2app
-
-To build:
-python3 -m venv env
-. venv/bin/activate
-rm -rf build dist
-python3 setup.py py2app -A
-
-Run from terminal (to see error message):
-./dist/DieStudy.app/Contents/MacOS/DieStudy
-"""
-
-from setuptools import setup
-import glob
-
-demo_files = []
-for folder in glob.glob('demo/*'):
-    demo_files.append( (folder, glob.glob(folder + "/*.*")) )
-
-APP = ['DieStudy.py']
-DATA_FILES = [
-    'data.json',
-    ('images', glob.glob('images/*.*'))
-    ]
-
-DATA_FILES += demo_files
-
-OPTIONS = {'packages': ['pandas']}
-
-setup(
-    app=APP,
-    data_files=DATA_FILES,
-    options={'py2app': OPTIONS},
-    setup_requires=['py2app'],
-)
